+++ src/display/utils/LineDrawer.hpp
@@ -9,24 +9,24 @@ class LineDrawer
 public:
     LineDrawer(const DynamicData &dynamicData) : dynamicData(dynamicData) {}
 
-    void drawTriggerIndicator(const int x, const int y);
+    void drawTriggerIndicator(const float x, const float y);
     void drawGrid(const int numOfVerticalLayers, const int numOfHorizontalLayers);
     void drawDisplayAreaBorder();
     void drawVerticalMeasurement();
     void drawHorizontalMeasurement();
 
 private:
-    void drawHorizontalLine(const int y, const float boldness,
-                            const bool stipple_line = false, const int x1 = 0,
-                            const int x2 = X_DISPLAY_RESOLUTION);
-    void drawVerticalLine(const int x, const float boldness,
-                          const bool stipple_line = false, const int y1 = 0,
-                          const int y2 = Y_DISPLAY_RESOLUTION);
-    void drawHorizontalLineWithLabels(const int y, const char *unit_label,
+    void drawHorizontalLine(const float y, const float boldness,
+                            const bool stipple_line = false, const float x1 = 0,
+                            const float x2 = X_DISPLAY_RESOLUTION);
+    void drawVerticalLine(const float x, const float boldness,
+                          const bool stipple_line = false, const float y1 = 0,
+                          const float y2 = Y_DISPLAY_RESOLUTION);
+    void drawHorizontalLineWithLabels(const float y, const char *unit_label,
                                       const TextPrinter &textPrinter,
                                       const float boldness_line = NEUTRAL,
                                       const bool stipple_line = false);
-    void drawVerticalLineWithLabels(const int x, const char *value_label,
+    void drawVerticalLineWithLabels(const float x, const char *value_label,
                                     const char *unit_label,
                                     const TextPrinter &textPrinter,
                                     const float boldness_line = NEUTRAL,
@@ -34,11 +34,12 @@ private:
     void drawHorizontalGrid(const int numOfHorizontalLayers);
     void drawVerticalGrid(const int numOfVerticalLayers);
 
-    TextPrinter textPrinterForTriggerIndicator{FONT_SIZE_MEDIUM, MONOCHROME_TEXT_COLOR_WHITE,
-                                               NEUTRAL};
-    TextPrinter textPrinterForGrid{FONT_SIZE_SMALL, MONOCHROME_TEXT_COLOR_DARK_GRAY, NEUTRAL};
-    TextPrinter textPrinterForMeasurements{FONT_SIZE_MEDIUM, MONOCHROME_TEXT_COLOR_WHITE,
-                                           NEUTRAL};
+    TextPrinter textPrinterForTriggerIndicator{
+        FONT_SIZE_MEDIUM, MONOCHROME_TEXT_COLOR_WHITE, NEUTRAL};
+    TextPrinter textPrinterForGrid{FONT_SIZE_SMALL,
+                                   MONOCHROME_TEXT_COLOR_DARK_GRAY, NEUTRAL};
+    TextPrinter textPrinterForMeasurements{FONT_SIZE_MEDIUM,
+                                           MONOCHROME_TEXT_COLOR_WHITE, NEUTRAL};
 
     const DynamicData &dynamicData;
 };
